refactor: moved batching logic from dataloader service to user and post services
